requirements.txt
gspread>=5.0.0
google-auth>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.18.0
requests>=2.28.0
